seperate models from queries
