+++ test/server.test.js
@@ -49,7 +49,7 @@ describe('POST /add', () => {
       response.body.should.have.property('response_type')
       response.body.response_type.should.equal('in_channel')
       response.body.should.have.property('text')
-      response.body.text.should.equal('<@U5GFS4CAE|keji> recieved a win')
+      response.body.text.should.equal('*<@U5GFS4CAE|keji> recieved a win*')
       done()
     })
   })
@@ -75,7 +75,7 @@ describe('POST /add', () => {
       response.body.should.have.property('response_type')
       response.body.response_type.should.equal('in_channel')
       response.body.should.have.property('text')
-      response.body.text.should.equal('<@U5GFS4CAE|keji> recieved a loss')
+      response.body.text.should.equal('*<@U5GFS4CAE|keji> recieved a loss*')
       done()
     })
   })
@@ -124,7 +124,7 @@ describe('POST /check', () => {
       response.should.have.status(200)
       response.body.should.be.a('object')
       response.body.should.have.property('text')
-      response.body.text.should.equal('You have 1 losses')
+      response.body.text.should.equal('*<@U5GFS4CAE|keji> has 6 losses*')
       done()
     })
   })
@@ -148,7 +148,7 @@ describe('POST /check', () => {
       response.should.have.status(200)
       response.body.should.be.a('object')
       response.body.should.have.property('text')
-      response.body.text.should.equal('You have 1 wins')
+      response.body.text.should.equal('*<@U5GFS4CAE|keji> has 6 wins*')
       done()
     })
   })
